kernel+libc: implement and use fstatat
This patch replaces stat, fstat, and lstat with fstatat at the library
level. fstatat handles the fstat case with the linux specific flag
AT_EMPTY_PATH, which the kernel now implements.

+++ makeimg.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
+if [ "$(id -u)" != 0 ];
+then
+    echo "this script must be ran as root"
+    exit 1
+fi
+
 dd if=/dev/zero of=os_2.img bs=516096c count=200
 
 losetup -o0 /dev/loop100 os_2.img
